+++ Tasks/Import/1.tasks.ps1
@@ -1,0 +1,9 @@
+
+# Tell to test the required variable.
+requires MyVar1
+
+# Synopsis: Some task with a variable.
+task MyVar {
+	# just show the value
+	"MyVar1 = $MyVar1"
+}