+++ Tasks/Import/2.tasks.ps1
@@ -1,0 +1,9 @@
+
+# Tell to test the required environment variable.
+requires -Environment MyEnv1
+
+# Synopsis: Some task with an environment variable.
+task MyEnv {
+	# just show the value
+	"MyEnv1 = $env:MyEnv1"
+}