+++ Tests/Property.test.ps1
@@ -32,7 +32,7 @@ equals $MissingNullProperty 42
 # Error: missing property
 task MissingProperty {
 	($r = try {property _111126_181750} catch {$_})
-	assert (($r | Out-String) -like "*Missing variable '_111126_181750'.*At *\Property.test.ps1:*ObjectNotFound*")
+	assert (($r | Out-String) -like "*Missing property '_111126_181750'.*At *\Property.test.ps1:*ObjectNotFound*")
 }
 
 # v3.3.4 (#60): Treat '' as not defined.